backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
langchain==0.1.9
langchain-community==0.0.24
docx2txt==0.8
python-docx==1.1.0
Pillow==10.2.0
pytesseract==0.3.10
sentence-transformers==2.5.1
pypdf==4.0.2 